placeholder, cleaned code, added Form component
